Download Start Problem fixed download loader exception fixed download itemcontrol template progressbar fixed

+++ HyperDownloadManager/Utils/IOUtility.cs
@@ -102,15 +102,21 @@ namespace HyperDownloadManager.Utils
                 return null;
             }
         }
-        public static BitmapSource GetFileIcon(string fileName)
+        public static BitmapSource? GetFileIcon(string fileName)
         {
-            Icon icon = null;
-            string path = Path.Combine(SpecialDirectories.Temp, fileName);
-            File.Create(path).Close();
-            icon = Icon.ExtractAssociatedIcon(path);
-            File.Delete(path);
-            BitmapSource bitmap = Imaging.CreateBitmapSourceFromHIcon(icon.Handle, Int32Rect.Empty, BitmapSizeOptions.FromEmptyOptions());
-            return bitmap;
+            if(fileName != "")
+            {
+                Icon? icon = null;
+                string path = Path.Combine(SpecialDirectories.Temp, fileName);
+                File.Create(path).Close();
+                icon = Icon.ExtractAssociatedIcon(path);
+                File.Delete(path);
+                if (icon == null)
+                    return null;
+                BitmapSource bitmap = Imaging.CreateBitmapSourceFromHIcon(icon.Handle, Int32Rect.Empty, BitmapSizeOptions.FromEmptyOptions());
+                return bitmap;
+            }
+            return null;
         }
         public static void OpenExplorer(string? fileSavePath)
         {
